+++ database/sharding/sharding.go
@@ -67,7 +67,7 @@ func (m *Manager) GetDB(key uint64) *gorm.DB {
 	m.mu.RLock()
 	defer m.mu.RUnlock()
 
-	shardIndex := int(key % uint64(m.shardCount)) // 经过审计，此处忽略是安全的。
+	shardIndex := int(key % uint64(uint32(m.shardCount)))
 
 	return m.shards[shardIndex].RawDB()
 }
